[SonTT][update list, delete employee ]

+++ src/main/java/com/example/retro_care/employee/controller/EmployeeController.java
@@ -56,7 +56,7 @@ public class EmployeeController {
         System.out.println("employeeDto");
         new EmployeeDto().validate(employeeDto, bindingResult);
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>(bindingResult.getAllErrors().toString(),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(bindingResult.getAllErrors().toString(), HttpStatus.BAD_REQUEST);
         }
         Employee employee = new Employee();
         BeanUtils.copyProperties(employeeDto, employee);
@@ -68,19 +68,20 @@ public class EmployeeController {
      * Author: TanNV
      * Date: 15/09/2023
      * Use to get employee by id and return Http status OK if it can find it else  return http status NO_CONTENT
+     *
      * @param id
      * @return Reponse entity
      */
     @GetMapping("/{id}")
-    public ResponseEntity<Employee> getEmployee(@PathVariable Long id){
-        if(id==null){
+    public ResponseEntity<Employee> getEmployee(@PathVariable Long id) {
+        if (id == null) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         Employee employee = employeeService.getById(id);
-        if(employee==null){
+        if (employee == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(employee,HttpStatus.OK);
+        return new ResponseEntity<>(employee, HttpStatus.OK);
     }
 
     /**
@@ -88,47 +89,27 @@ public class EmployeeController {
      * Date: 16/09/2023
      * Receive data and validate, if there is an error, return BAD_REQUEST,
      * then save the employee to the DB. If saved successfully, return OK
-     * @param id id employee
-     * @param employeeDto validate info
+     *
+     * @param id            id employee
+     * @param employeeDto   validate info
      * @param bindingResult return error
      * @return Responese Entity with message
      */
     @PatchMapping("/update/{id}")
     public ResponseEntity<String> updateEmployee(@PathVariable Long id,
-                                                   @RequestBody EmployeeDto employeeDto,
-                                                   BindingResult bindingResult){
+                                                 @RequestBody EmployeeDto employeeDto,
+                                                 BindingResult bindingResult) {
         new EmployeeDto().validate(employeeDto, bindingResult);
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>(bindingResult.getAllErrors().toString(),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(bindingResult.getAllErrors().toString(), HttpStatus.BAD_REQUEST);
         }
         Employee employee = employeeService.getById(id);
-        if(employee==null){
-            return new ResponseEntity<>("Không tìm thấy",HttpStatus.NOT_FOUND);
+        if (employee == null) {
+            return new ResponseEntity<>("Không tìm thấy", HttpStatus.NOT_FOUND);
         }
         BeanUtils.copyProperties(employeeDto, employee);
         employeeService.updateEmployee(employee);
-        return new ResponseEntity<>("Update successfully",HttpStatus.OK);
-    }
-    /**
-     * Create: SonTT
-     * Date create: 15/09/2023
-     * Function: Call the database to retrieve the data with page, limit and sort
-     *
-     * @param page
-     * @param limit
-     * @return ResponseEntity<?>
-     */
-    @GetMapping("/get-list/{page}/{limit}/{sort}")
-    public ResponseEntity<Page<Employee>> getListEmployee(@PathVariable(value = "page", required = false) Integer page,
-                                                          @PathVariable(value = "limit", required = false) Integer limit,
-                                                          @PathVariable(value = "sort", required = false) String sort) {
-        Pageable pageable = PageRequest.of(page, limit, Sort.by(Sort.Direction.ASC, sort));
-        Page<Employee> employees = employeeService.getListEmployee(pageable);
-        if (employees.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        } else {
-            return new ResponseEntity<>(employees, HttpStatus.OK);
-        }
+        return new ResponseEntity<>("Update successfully", HttpStatus.OK);
     }
 
 
@@ -136,43 +117,58 @@ public class EmployeeController {
      * Create: SonTT
      * Date create: 15/09/2023
      * Function: Call the database to retrieve paginated data with fields idRole and employee name
+     *
      * @param page
      * @param limit
      * @param sort
-     * @param idRole
      * @param nameEmployee
      * @return ResponseEntity<?>
      */
-    @GetMapping("/search-list/{page}/{limit}/{sort}")
+    @GetMapping("/list/{page}/{limit}/{sort}")
     public ResponseEntity<Page<Employee>> searchEmployee(@PathVariable(value = "page", required = false) Integer page,
                                                          @PathVariable(value = "limit", required = false) Integer limit,
                                                          @PathVariable(value = "sort", required = false) String sort,
-                                                         @RequestParam(value = "role", required = false) Long idRole,
                                                          @RequestParam(value = "name", required = false) String nameEmployee) {
-        Pageable pageable = PageRequest.of(page, limit,  Sort.by(Sort.Direction.ASC, sort));
-        Page<Employee> employees = employeeService.searchEmployee(pageable, idRole, nameEmployee);
-        if (employees.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        } else {
-            return new ResponseEntity<>(employees, HttpStatus.OK);
-        }
+        Pageable pageable = PageRequest.of(page, limit, Sort.by(Sort.Direction.ASC, sort));
+        Page<Employee> employees;
+        if (nameEmployee == null) {
+             employees = employeeService.getListEmployee(pageable);
+            if (employees.isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(employees, HttpStatus.OK);
+            }
 
+        } else {
+             employees = employeeService.searchEmployee(pageable, nameEmployee);
+            if (employees.isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(employees, HttpStatus.OK);
+            }
+        }
     }
 
     /**
      * Create:SonTT
      * Date create: 15/09/2023
      * Function: with the correct input parameter id true then return HttpStatus.OK otherwise return false
+     *
      * @param id
      * @return ResponseEntity<>
      */
     @DeleteMapping("/delete-employee")
     public ResponseEntity<HttpStatus> deleteEmployee(@RequestParam(value = "id", required = false) Long id) {
-        if (employeeService.deleteEmployee(id)) {
-            return new ResponseEntity<>(HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (employeeService.findEmployee(id)==null){
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }else {
+            if (employeeService.deleteEmployee(id)) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
+
 
     }
